Support running on OSX for development
To get this to start on my 2017 MBP running Mojave, I needed to
recognize x86_64 hardware and tell it to not have managed networking. I
suppose the latter is a command line option, but there was a case for
Windows being able to continue without using it so I went with that
convention.

```
$ uname -a
Darwin seans-mbp-3.lan 18.7.0 Darwin Kernel Version 18.7.0: Thu Jan 23 06:52:12 PST 2020; root:xnu-4903.278.25~1/RELEASE_X86_64 x86_64
```