Fail quickly if essential items don't work
There are some cases where Main calls `return`, presumably with the
intention of exiting, but due to the NT and ScriptRunner threads
running, the program will never exit. Also, at least on my machine,
having no USB cameras throws an exception that's not caught, leaving the
app in limbo.

This change forces all threads to stop if no cameras or vision pipelines
can be instantiated. This will let the supervisor (e.g. systemd) restart
the app to handle conditions like cameras being unplugged prior to the
boot.

During our season we found that moving the camera port would cause this
exception because the USB ID changed. We would have to then go hard
cycle the Pi, when it would have been much easier to just move the port
and know that CV would restart shortly.